backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2
scipy==1.11.4
opencv-python==4.8.1.78
requests==2.31.0
grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==4.25.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
cors==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
scansegmentapi @ git+https://github.com/SICKAG/ScanSegmentAPI
open3d==0.18.0
